move collections to `collections` folder
