+++ src/app/collections/Catalog.ts
@@ -1,4 +1,4 @@
-import type Book from './Book';
+import type Book from '../models/Book';
 
 export default class Catalog {
   private _products: Book[];
